Remove excess lines
