+++ src/test/java/seedu/address/logic/parser/tagparsers/TagDeleteCommandParserTest.java
@@ -1,4 +1,4 @@
-package seedu.address.logic.parser;
+package seedu.address.logic.parser.tagparsers;
 
 import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 import static seedu.address.logic.parser.CommandParserTestUtil.assertParseFailure;
@@ -7,28 +7,27 @@ import static seedu.address.testutil.TypicalIndexes.INDEX_FIRST_TAG;
 
 import org.junit.jupiter.api.Test;
 
-import seedu.address.logic.commands.tagcommands.LinkDeleteCommand;
-import seedu.address.logic.parser.tagparsers.LinkDeleteCommandParser;
+import seedu.address.logic.commands.tagcommands.TagDeleteCommand;
+import seedu.address.logic.parser.tagparsers.TagDeleteCommandParser;
 
 /**
  * As we are only doing white-box testing, our test cases do not cover path variations
- * outside of the LinkDeleteCommand code. For example, inputs "1" and "1 abc" take the
- * same path through the LinkDeleteCommand, and therefore we test only one of them.
+ * outside of the TagDeleteCommand code. For example, inputs "1" and "1 abc" take the
+ * same path through the TagDeleteCommand, and therefore we test only one of them.
  * The path variation for those two cases occur inside the ParserUtil, and
  * therefore should be covered by the ParserUtilTest.
  */
-public class LinkDeleteCommandParserTest {
-
-    private LinkDeleteCommandParser parser = new LinkDeleteCommandParser();
+public class TagDeleteCommandParserTest {
+    private TagDeleteCommandParser parser = new TagDeleteCommandParser();
 
     @Test
-    public void parse_validArgs_returnsLinkDeleteCommand() {
-        assertParseSuccess(parser, "1", new LinkDeleteCommand(INDEX_FIRST_TAG));
+    public void parse_validArgs_returnsTagDeleteCommand() {
+        assertParseSuccess(parser, "1", new TagDeleteCommand(INDEX_FIRST_TAG));
     }
 
     @Test
     public void parse_invalidArgs_throwsParseException() {
         assertParseFailure(parser, "a",
-                String.format(MESSAGE_INVALID_COMMAND_FORMAT, LinkDeleteCommand.MESSAGE_USAGE));
+                String.format(MESSAGE_INVALID_COMMAND_FORMAT, TagDeleteCommand.MESSAGE_USAGE));
     }
 }
